Reorganize binary search tree model
